[DEV-9734b] rename and add TS
