+++ packages/chart/src/hooks/useIntersectionObserver.ts
@@ -1,0 +1,37 @@
+import { useEffect, useState, MutableRefObject } from 'react'
+
+interface IntersectionObserverOptions {
+  threshold?: number | number[]
+  root?: Element | null
+  rootMargin?: string
+  freezeOnceVisible?: boolean
+}
+
+export default function useIntersectionObserver(
+  elementRef: MutableRefObject<Element | null>,
+  { threshold = 0, root = null, rootMargin = '0%', freezeOnceVisible = false }: IntersectionObserverOptions
+) {
+  const [entry, setEntry] = useState<IntersectionObserverEntry | undefined>()
+
+  const frozen = entry?.isIntersecting && freezeOnceVisible
+
+  const updateEntry = ([entry]: IntersectionObserverEntry[]) => {
+    setEntry(entry)
+  }
+
+  useEffect(() => {
+    const node = elementRef?.current
+    const hasIOSupport = !!window.IntersectionObserver
+
+    if (!hasIOSupport || frozen || !node) return
+
+    const observerParams = { threshold, root, rootMargin }
+    const observer = new IntersectionObserver(updateEntry, observerParams)
+
+    observer.observe(node)
+
+    return () => observer.disconnect()
+  }, [elementRef, threshold, root, rootMargin, frozen])
+
+  return entry
+}
